html/browsers/browsing-the-web/back-forward-cache/eligibility/inflight-fetch.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="../resources/helper.sub.js"></script>
<script>
promise_test(async t => {
  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  window.open(urlA, '_blank', 'noopener');

  await pageA.execute_script(waitForPageShow);

  await pageA.execute_script(
    (url) => {
      window.fetchPromise = fetch('/common/slow.py');
      prepareNavigation(() => { location.href = url; });
    },
    [urlB]);

  await pageB.execute_script(waitForPageShow);
  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );
  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  assert_equals(
    await pageA.execute_script(() => window.fetchPromise.then(r => r.text())),
    '',
    'Fetch should complete');
}, 'Eligibility (in-flight fetch)');
</script>
